Copy ipaddr library to be used for more configurable toolbar restricting.
I looked at the ipaddr history:

http://lwn.net/Articles/337684/

Given that ipaddr is not already 2 and 3 compatible while netaddr is
compatible and is available as a dependency, I evaluated netaddr.  I
found it to be a little big, and having some unconventional stuff in
it.

Then I tried copying ipaddr from the URL below and making it
compatible with both 2 and 3 which looks pretty easy so I'll just take
that route:

http://code.google.com/p/ipaddr-py/source/browse/branches/2.1.x/?r=a86bf59c63cf8f1d59282e538cc5f5c3e650dbfa