Fix issue preventing console use with Google App Engine
A merge of recent changes to werkzeug's debug console.  This patch
addresses an issue with the console and Google App Engine which was
preventing the result of code operations from reaching the actual
console.  With this patch sys.stdout is saved to a temporary variable
for restoration after the code is run.