Add new debugtoolbar. settings for configuration
Since pyramid_debugtoolbar is a stand-alone application from the view
point of Pyramid, it does not inherit any of the settings from the
application it was started from. This means that things like
pyramid.reload_templates and pyramid.debug_notfound are not brought over
from the end-users application.

In most cases this is desirable, it becomes problematic when the user is
attempting to create a debugtoolbar panel and is constantly modifying
the templates.

This changes introduces a bunch of new configuration variables that may
be set, they are outside the scope of the standard Pyramid configuration
and are namespaced to the debugtoolbar. This allows one for instance to
turn on the reload_templates feature in the debugtoolbar, while keeping
it turned off in the end-users application.