monkeypatch the pyramid router to use as a BeforeApplicationCreated hook
Pyramid does not have a BeforeApplicationCreated event. In general this
is good because it would compete with the excellent conflict resolution
system. However, the Configurator supports multiple commit cycles and so
it's very difficult to apply configuration actions **after** all user
code is done so that we can introspect and wrap the final config. This
is only really necessary for something like the toolbar that is trying
to play along without being noticed. To accomplish this we monkeypatch
``pyramid.config.Router`` which is used to create the router object when
``config.make_wsgi_app`` is invoked. From there we apply our final
actions and then return the original router object.

A previous iteration of this used the ``ApplicationCreated`` event but
the router uses the registry immediately - thus many actions, like
swapping the root factory, would not be noticed. We must apply our
actions before the router object is actually created.