Remove parsing of REMOTE_ADDR when it contains a comma
Some middleware box, not identified in issue #85 is inserting additional
IP addresses for the proxies in the REMOTE_ADDR environ variable.

That's not what it is for, it is supposed to provide the IP address of
the client that made the original request, and then X-Forwarded-For
contains additional information.

However middleware boxes being the type of quality they usually are, we
can't rely on them always placing the proxy IP address first, followed
by the client IP address last.

Especially since X-Forwarded-For is specified as <client>, <proxy 1>,
<proxy 2> and thus is the reverse of whateve this mess is.

So instead we bail out. If you own such a middleware box that sets
REMOTE_ADDR this way, please, for all that is good, please let me know
what it is and who authored it.

Your final WSGI server should parse X-Forwarded-For/Forwarded (if
received from a trusted proxy) and place the client IP in REMOTE_ADDR.

It should then also pass on the X-Forwarded-For/Forwarded header to the
application, as normal.